old-tests/submission/W3C/domhtml/HTMLImageElement03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IMG</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-img-element.html#dom-img-alt" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<img align="middle" alt="DTS IMAGE LOGO" border="0" height="50" hspace="4" id="IMAGE-1" longdesc="./files/desc.html" name="IMAGE-1" src="/support/images/yellow-rectangle.png" usemap="#DTS-MAP" vspace="10" width="100">
</p>



<script>
var domtest = async_test("The alt attribute specifies an alternative text for user agenst not     rendering the normal content of this element.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var valt;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("img");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      valt = testNode.alt;

      assert_equals(valt,"DTS IMAGE LOGO");
       
    domtest.done();
}));

</script>
<p>The alt attribute specifies an alternative text for user agenst not
    rendering the normal content of this element.

    Retrieve the alt attribute and examine its value.</p>
<div id="log"></div>
</body></html>
